% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYear.R
\name{getYear}
\alias{getYear}
\title{GetYear
Identify year of publishing of movie}
\usage{
getYear(title)
}
\arguments{
\item{title}{character}
}
\value{
numeric
}
\description{
GetYear
Identify year of publishing of movie
}
