% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMovie.R
\name{plotMovie}
\alias{plotMovie}
\title{Plots Movies Length vs rating and highlights selected movie}
\usage{
plotMovie(queryTitle, color = "green")
}
\arguments{
\item{queryTitle}{character}

\item{color}{character}
}
\value{
plot
}
\description{
Plots Movies Length vs rating and highlights selected movie
}
